Inline interface for SqlStandardAccessControlMetastore
